+++ src/main/java/seedu/meetingjio/common/ErrorMessages.java
@@ -13,6 +13,7 @@ public class ErrorMessages {
             + "st/[StartTime]\n"
             + "et/[EndTime]\n"
             + "m/[Mode]";
+
     public static final String ERROR_MISSING_VALUES = "Hi 1 or more parameters have missing values\n"
             + "The parameters that require values are:\n"
             + "n/[Name]\n"
@@ -61,4 +62,29 @@ public class ErrorMessages {
             + "If there is no such constraint for the duration, just input 'free'.";
     public static final String ERROR_NO_FREE_TIMESLOT = "Sorry, there is no timeslot that fits your requirement.";
     public static final String ERROR_UNSPECIFIED_LIST_CLEAR = "Please specify which timetable to Clear";
+    public static final String ERROR_MISSING_PARAMETERS_DELETE = "Hi 1 or more parameters are missing\n"
+            + "The required parameters are:\n"
+            + "n/[Name]\n"
+            + "i/[INDEX]\n";
+    public static final String ERROR_MISSING_VALUES_DELETE = "Hi 1 or more parameters are missing\n"
+            + "The required parameters are:\n"
+            + "n/[Name]\n"
+            + "i/[INDEX]\n";
+
+    public static final String ERROR_MISSING_PARAMETERS_ADD_MEETING = "Hi 1 or more parameters are missing\n"
+            + "The required parameters are:\n"
+            + "t/[Title]\n"
+            + "d/[Day]\n"
+            + "st/[StartTime]\n"
+            + "et/[EndTime]\n"
+            + "m/[Mode]";
+
+    public static final String ERROR_MISSING_VALUES_ADD_MEETING = "Hi 1 or more parameters have missing values\n"
+            + "The parameters that require values are:\n"
+            + "t/[Title]\n"
+            + "d/[Day]\n"
+            + "st/[StartTime]\n"
+            + "et/[EndTime]\n"
+            + "m/[Mode]";
+
 }
