+++ src/main/java/seedu/meetingjio/parser/ParserHelperMethods.java
@@ -1,0 +1,77 @@
+package seedu.meetingjio.parser;
+
+import seedu.meetingjio.exceptions.InvalidDayException;
+import seedu.meetingjio.exceptions.InvalidModeException;
+import seedu.meetingjio.exceptions.InvalidTimeException;
+import seedu.meetingjio.exceptions.MissingValueException;
+
+public class ParserHelperMethods {
+
+    /**
+     * Checks the validity of the user's given startTime and endTime.
+     *
+     * @param startTime Time at which the event begins
+     * @param endTime   Time at which the event ends
+     * @throws InvalidTimeException If the given hours and minutes are invalid, or if startTime is later than endTime
+     */
+    void checkTime(int startTime, int endTime) throws InvalidTimeException {
+        int startMinutes = startTime % 100;
+        int endMinutes = endTime % 100;
+        boolean invalidMinutes = startMinutes >= 60 || endMinutes >= 60;
+        int startHours = startTime / 100;
+        int endHours = endTime / 100;
+        boolean invalidHours = startHours >= 24 || endHours >= 24;
+        if (invalidMinutes || invalidHours || startTime > endTime) {
+            throw new InvalidTimeException();
+        }
+    }
+
+    /**
+     * Checks that all parameters have a non-null value.
+     *
+     * @param eventDescription Array of user's input
+     * @throws MissingValueException If at least one parameter has no value
+     */
+    protected void checkNonNullValues(String[] eventDescription,int lastElementIndex) throws MissingValueException {
+        for (int i = 0; i < lastElementIndex; i++) {
+            if (eventDescription[i].length() == 0) {
+                throw new MissingValueException();
+            }
+            assert (eventDescription[i].length() != 0) : "The parameters have non-null values";
+        }
+    }
+
+    /**
+     * Checks that the mode given by the user is either online or physical.
+     *
+     * @param mode String given by user
+     * @throws InvalidModeException If mode is neither online nor physical
+     */
+    protected void checkMode(String mode) throws InvalidModeException {
+        if (mode.equals("online") || mode.equals("physical")) {
+            return;
+        }
+        throw new InvalidModeException();
+    }
+
+    /**
+     * Ensures that the 'day' parameter in user's input is a valid day.
+     *
+     * @param day String given by user
+     * @throws InvalidDayException If value of 'day' does not correspond to an actual day
+     */
+    void checkDay(String day) throws InvalidDayException {
+        switch (day) {
+        case "monday":
+        case "tuesday":
+        case "wednesday":
+        case "thursday":
+        case "friday":
+        case "saturday":
+        case "sunday":
+            break;
+        default:
+            throw new InvalidDayException();
+        }
+    }
+}
